I did Crud for course prices

+++ database/migrations/2022_09_26_130943_create_course_prices_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('course_prices', function (Blueprint $table) {
             $table->id();
-            $table->integer('cost');
+            $table->bigInteger('cost');
             $table->foreignId('department_id')->constrained()->onDelete('cascade');
             $table->foreignId('course_id')->constrained()->onDelete('cascade');
             $table->timestamps();
